separate pricing components into its own files
